refactor(profiling): simplify string tables
