+++ profiling/src/profiling/stalk_walking.rs
@@ -1,7 +1,7 @@
 use crate::bindings::{
     ddog_php_prof_zend_string_view, zend_execute_data, zend_function, ZEND_USER_FUNCTION,
 };
-use crate::string_table::{OwnedStringTable, StringTable};
+use crate::string_table::StringTable;
 use std::borrow::Cow;
 use std::cell::{RefCell, RefMut};
 use std::mem::transmute;
@@ -28,7 +28,7 @@ impl FunctionRunTimeCacheStats {
 
 #[cfg(php_run_time_cache)]
 thread_local! {
-    static CACHED_STRINGS: RefCell<OwnedStringTable> = RefCell::new(OwnedStringTable::new());
+    static CACHED_STRINGS: RefCell<StringTable> = RefCell::new(StringTable::new());
     pub static FUNCTION_CACHE_STATS: RefCell<FunctionRunTimeCacheStats> = RefCell::new(Default::default())
 }
 
@@ -37,7 +37,7 @@ thread_local! {
 #[inline]
 pub unsafe fn activate_run_time_cache() {
     #[cfg(php_run_time_cache)]
-    CACHED_STRINGS.with(|cell| cell.replace(OwnedStringTable::new()));
+    CACHED_STRINGS.with(|cell| cell.replace(StringTable::new()));
 }
 
 #[derive(Default, Debug)]
@@ -92,11 +92,11 @@ unsafe fn extract_function_name(func: &zend_function) -> Option<String> {
 
 unsafe fn handle_file_cache_slot_helper(
     execute_data: &zend_execute_data,
-    string_table: &mut RefMut<OwnedStringTable>,
+    string_table: &mut RefMut<StringTable>,
     cache_slots: &mut [usize; 2],
 ) -> Option<Cow<'static, str>> {
     let offset = if cache_slots[1] > 0 {
-        cache_slots[1]
+        cache_slots[1] as u32
     } else {
         // Safety: if we have cache slots, we definitely have a func.
         let func = &*execute_data.func;
@@ -106,7 +106,7 @@ unsafe fn handle_file_cache_slot_helper(
             return None;
         };
         let offset = string_table.insert(file.as_ref());
-        cache_slots[1] = offset;
+        cache_slots[1] = offset as usize;
         offset
     };
     let str = string_table.get_offset(offset);
@@ -120,7 +120,7 @@ unsafe fn handle_file_cache_slot_helper(
 
 unsafe fn handle_file_cache_slot(
     execute_data: &zend_execute_data,
-    string_table: &mut RefMut<OwnedStringTable>,
+    string_table: &mut RefMut<StringTable>,
     cache_slots: &mut [usize; 2],
 ) -> (Option<Cow<'static, str>>, u32) {
     match handle_file_cache_slot_helper(execute_data, string_table, cache_slots) {
@@ -137,17 +137,17 @@ unsafe fn handle_file_cache_slot(
 
 unsafe fn handle_function_cache_slot(
     func: &zend_function,
-    string_table: &mut RefMut<OwnedStringTable>,
+    string_table: &mut RefMut<StringTable>,
     cache_slots: &mut [usize; 2],
 ) -> Option<Cow<'static, str>> {
     if cache_slots[0] > 0 {
-        let offset = cache_slots[0];
+        let offset = cache_slots[0] as u32;
         let str = string_table.get_offset(offset);
         Some(Cow::Owned(str.to_string()))
     } else {
         let name = extract_function_name(func)?;
         let offset = string_table.insert(name.as_ref());
-        cache_slots[0] = offset;
+        cache_slots[0] = offset as usize;
         Some(Cow::Owned(name))
     }
 }
